refactor(cli): give the shared CLI helpers a home outside main.py
First step of splitting `cli/main.py` (7,730 lines) into per-group modules.
`main.py` will import each command group from `cli/commands/`, so any helper a
group module also needs cannot keep living in `main.py`: importing it back
would make the dependency cyclic. Move the two genuinely cross-group helpers
into a new `cli/_shared.py`, which both sides import downwards:

* `_activate_s3` - 48 call sites, spread across most groups
* `create_default_group` - the factory behind the `check`, `extract` and
  `inspect` default-subcommand groups

`cli/decorators.py` stays the home for the reusable Click option decorators and
the `cls=` command classes (`GlobAwareCommand`, `SingleFileCommand`,
`parse_row_group_options`); nothing was moved out of it, and `_shared.py`
deliberately holds only the group-neutral runtime plumbing that is not a
decorator. Helpers used by a single group (`OptionalIntCommand`,
`ConvertDefaultGroup`, the three generated `*DefaultGroup` classes) stay in
`main.py` and will travel with their group.

Code is moved verbatim. `main.py` re-imports both names, so every existing
call site and `from geoparquet_io.cli.main import ...` keeps working.
`tests/data/cli_surface.json` is unchanged.
