+++ geoparquet_io/cli/main.py
@@ -1,6 +1,5 @@
 import os
 import sys
-from contextlib import contextmanager
 from importlib.metadata import entry_points
 from pathlib import Path
 
@@ -9,6 +8,7 @@ import duckdb
 from click.core import ParameterSource
 from click_plugins import with_plugins
 
+from geoparquet_io.cli._shared import _activate_s3, create_default_group
 from geoparquet_io.cli.decorators import (
     GlobAwareCommand,
     SingleFileCommand,
@@ -133,34 +133,6 @@ class OptionalIntCommand(GlobAwareCommand):
         return super().make_context(info_name, args, parent=parent, **extra)
 
 
-@contextmanager
-def _activate_s3(ctx, aws_profile=None, s3_endpoint=None, s3_region=None, s3_no_ssl=False):
-    """Resolve and activate S3 config from ctx + per-command overrides.
-
-    Properly saves/restores AWS_PROFILE env var to avoid credential leaks.
-    """
-    from geoparquet_io.core.duckdb_utils import s3_config_scope
-    from geoparquet_io.core.remote import resolve_s3_config
-
-    config = resolve_s3_config(
-        s3_endpoint=s3_endpoint or ctx.obj.get("s3_endpoint"),
-        s3_region=s3_region or ctx.obj.get("s3_region"),
-        s3_no_ssl=s3_no_ssl or ctx.obj.get("s3_no_ssl", False),
-        aws_profile=aws_profile or ctx.obj.get("aws_profile"),
-    )
-    previous_profile = os.environ.get("AWS_PROFILE")
-    try:
-        if config["profile"]:
-            os.environ["AWS_PROFILE"] = config["profile"]
-        with s3_config_scope(config):
-            yield config
-    finally:
-        if previous_profile is None:
-            os.environ.pop("AWS_PROFILE", None)
-        else:
-            os.environ["AWS_PROFILE"] = previous_profile
-
-
 @with_plugins(entry_points(group="gpio.plugins"))
 @click.group()
 @click.version_option(prog_name="geoparquet-io")
@@ -205,35 +177,6 @@ def cli(ctx, timestamps, s3_endpoint, s3_region, s3_no_ssl, aws_profile):
     ctx.obj["aws_profile"] = aws_profile
     # Setup logging for CLI output (default level INFO, verbose commands will set DEBUG)
     setup_cli_logging(verbose=False, show_timestamps=timestamps)
-
-
-# Check commands group - use custom command class for default subcommand
-def create_default_group(default_subcommand: str, description: str) -> type:
-    """Factory to create a click.Group subclass that defaults to a specific subcommand.
-
-    Args:
-        default_subcommand: The subcommand to invoke when none is provided
-        description: The docstring for the generated class
-
-    Returns:
-        A click.Group subclass with the configured default behavior
-    """
-
-    class _DefaultGroup(click.Group):
-        def parse_args(self, ctx, args):
-            # Handle --help for group
-            if "--help" in args and (not args or args[0] not in self.commands):
-                return super().parse_args(ctx, [a for a in args if a != "--help"] + ["--help"])
-
-            # If first arg is a known subcommand, use it
-            if args and not args[0].startswith("-") and args[0] in self.commands:
-                return super().parse_args(ctx, args)
-
-            # Default to configured subcommand
-            return super().parse_args(ctx, [default_subcommand] + args)
-
-    _DefaultGroup.__doc__ = description
-    return _DefaultGroup
 
 
 # Create default group classes using the factory
